commit-3 gestion compte

+++ fournisseurIdentiteApi/src/main/java/web/projet/fournisseurIdentite/repositories/UtilisateurRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import web.projet.fournisseurIdentite.models.Utilisateur;
 
 @Repository
-public interface UtilisateurRepository extends JpaRepository<Utilisateur,Integer> {
+public interface UtilisateurRepository extends JpaRepository<Utilisateur,Long> {
     Optional<Utilisateur> findByEmail(String email);
 }
